Moves modules into Project and Generator
