+++ waspc/cli/src/Wasp/Cli/Command/CreateNewProject/StarterTemplates/Templating.hs
@@ -9,9 +9,9 @@ import qualified Data.Text as T
 import StrongPath (Abs, Dir, File, Path')
 import Wasp.Cli.Command.CreateNewProject.Common (defaultWaspVersionBounds)
 import Wasp.Cli.Command.CreateNewProject.ProjectDescription (NewProjectAppName, NewProjectName)
-import Wasp.ExternalConfig.PackageJson (findPackageJsonFile)
 import Wasp.Project.Analyze (findWaspFile)
 import Wasp.Project.Common (WaspProjectDir)
+import Wasp.Project.ExternalConfig (findPackageJsonFile)
 import qualified Wasp.Util.IO as IOUtil
 
 replaceTemplatePlaceholdersInTemplateFiles :: NewProjectAppName -> NewProjectName -> Path' Abs (Dir WaspProjectDir) -> IO ()
