+++ waspc/src/Wasp/Generator/Common.hs
@@ -11,7 +11,6 @@ module Wasp.Generator.Common
     GeneratedSrcDir,
     makeJsArrayFromHaskellList,
     dropExtensionFromImportPath,
-    reactRouterVersion,
   )
 where
 
@@ -57,9 +56,6 @@ prismaVersion :: SV.Version
 -- NOTE: If changing prisma version here, also change it in waspc/packages/prisma/package.json.
 --       Then, make sure `data/Generator/templates/sdk/wasp/prisma-runtime-library.d.ts` is up to date.
 prismaVersion = SV.Version 5 19 1
-
-reactRouterVersion :: SV.ComparatorSet
-reactRouterVersion = SV.backwardsCompatibleWith $ SV.Version 5 3 3
 
 makeJsonWithEntityData :: String -> Aeson.Value
 makeJsonWithEntityData name =
